move tests to top level directory
this makes it easier to omit tests in the published package
