+++ packages/react-components/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'jsdom',
-  collectCoverageFrom: ['lib/**/*.{js,jsx,ts,tsx}', '!lib/**/tests/*'],
+  collectCoverageFrom: ['lib/**/*.{js,jsx,ts,tsx}'],
 };
