+++ packages/react-components/tests/commands/delivery-request-form.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import React from 'react';
-import { DeliveryRequestForm } from '..';
+import { DeliveryRequestForm } from '../../lib';
 import { availableDispensers, availablePlaces, fleets } from './test-data';
 
 describe('Form validation', () => {
