+++ packages/react-components/tests/lifts/lift-accordion.test.tsx
@@ -1,6 +1,6 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import { LiftAccordion } from '..';
+import { LiftAccordion } from '../../lib';
 import { makeLift } from './test-utils';
 
 test(`smoke test`, () => {
